Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Oh also. We could craft a job that just runs:
> 
>  mvn -DskipTests=true clean package
> 
> Which would at least give some kind of basic confidence. But I am not sure it is any helpful :/

The point is to run the unit tests, so -DskipTests=true setup would be useless. If it's currently failing, it's ok - we can work on it and see why it fails and amend the environment probably. I just have no idea how to do it yet.

Patch-set: 2
Reviewer: Smalyshev <2096@e9e9afe9-4712-486d-8885-f54b72dd1951>