Default selenium-daily-beta-{project} to node10/npm6
The container releng/npm-browser-test:0.4.0 is based on node6/npm3 and
we have to upgrade to node10/npm6 (T211784).
Also, some node dependencies require a full python2.7 version (T213944)
but the releng/npm-browser-test container has not been updated for it.

As a preparation step, switch selenium-daily-beta-{project} to the new
container shipping python2.7, node 10 and npm 6.

Make each job to explicitly use the legacy container. As a result this
change does not modify any jobs. They can then be switched individually.

Bug: T213944
Bug: T217545
Change-Id: I60d87b5129113963da396d551b9c38a28ba03aa9