Service Pipeline: correct imagename
ProductionImageID (what is currently used for the image name in
comments) is actually the Docker ID of the image being used -- this is
not useful for the intended audience of comments (end-users). We should
be using a fully qualified image name instead. This is easy to get from
the libpipeline library PipelineRunner class.

Change-Id: I2ee8ef6183457575346525469a380e39b6aa8ec5