Update patch set 3
Patch Set 3:

That uses Chromium 71, not 72. The reason it was broken is because I created a simpler fork of the quibble image without Firefox-ESR and discovered that without that firefox package, Chromium Headless does not work.

That is because Chromium has dependencies on libraries not packaged in the Debian package, and not marked as dependency by Debian. These libraries are only needed in Headless mode.

It has worked elsewhere by accident because firefox-esr happens to provide those libraries. It is unrelated, and has been fixed already by re-installing firefox-esr evnethough Fresnel does not need it.

Eventually, it should be fixed by Debian upstream, or here by listing all the packages manually from the troubleshooting.md link.

Patch-set: 3
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>