docker: sync node10-test-browser with npm-browser-test
The node10-test-browser can not be used to run the Selenium wdio tests:
we require chromedriver and Xvfb to be started. Add a few missing
packages and copy the entry point run-with-xvfb.sh

Add the same example-run.sh to quickly test an image. It uses
data-values/value-view which is a small repository, has a fast test
suite and is configured for Chromium, Firefox and PhantomJS. An
excellent smoke test.

npm-browser-test:
* Tweak a comment, not worth rebuilding an image for that.
* Change example-run.sh to be executable.

Bug: T217545
Change-Id: I5a33a7a9c74da60ce8902ed3de3749b23d7ed907