Update patch set 1
Patch Set 1:

@Hashar this is primarily used for npm-test jobs on non-mediawiki repositories. E.g. visualeditor, oojs, unicode, various performance/* repos, etc.

As for Selenium WDIO tests. I don't understand, those need MediaWiki, right? Would that not run on quibble instead?

As for Xvfb, I intentionally left that out to see if we still need it. As far as I know, we should not be using that anymore. We should be using headless right?

Thanks to not supporting Xvfb, I realised that quite a number of repos still used plain Chrome or Firefox instead of headless.

As for PhantomJS: This should not be used anymore. But, if any npm job still needs it, those legacy jobs probably already install it manually via npm (yay binaries). Not pretty, but I don't think we should boat the image for it.

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>