Update patch set 1
Patch Set 1:

If the are non-quibble-based selenium WDIO tests and they also need to record a video, then I suppose we can install Xvfb still. That would be fine I think. However, I do think we should in that case still keep the entry point like it is now. So that it is opt-in for the jobs that need it. The DISPLAY var would then be set from the run-with-xvfb.sh script.

It is also possible to launch Xvfb from the test jobs. The "xvfb-run" commands is available to do exactly that. It also starts an instance, waits for it, runs the command, and shuts down afterwards and sets up DISPLAY variable. For example "npm run do-daily" would be "xvfb-run npm run do-daily". I haven't tried this though, so run-with-xvfb.sh is fine too, but we could maybe use the --entrypoint flag from JJB to set it only when needed?

Patch-set: 1
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>