post pipeline: fix ref filter
In 45b176831e3423231f2048b795759b86da9dfc91 the "post" pipeline was
added saying, "post would react whenever a reference is updated
(refs/*)"; however, the definition of the post pipeline uses a negative
lookahead to filter any changes to "refs/.*". This seems like the
opposite of what this is supposed to do.

It seems that train-deploy-notes are the only thing currently using the
post pipeline. This job was not triggered when the recent
wmf/1.33.0-wmf.20 branch was pushed for mediawiki/core which led to the
creation of this patch.

Change-Id: Ib71a3b3a843b3a1113f267c9c50ddea4e77469be