Update patch set 1
Patch Set 1: Code-Review+2

From a discussion with cscott: 

php7.0 is EOL. Zest will be used for Parsoid port to PHP which would be merged in mediawiki/core in a couple years or so. So there is no point in supporting 7.0 / 7.1.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2