Drop erroneous 'goals' variable for maven jobs
wikidata-query-rdf and wikidata-query-blazegraph use the template
'{name}-maven-java8-docker' and both pass:

 goals: clean verify

However the template does not expand "goals" at all, that should be
"maven_args" which is passed as an argument to the java8/maven Docker
container.
The default is already "clean verify". Hence just entirely remove the
unused goals variable.

Change-Id: I2a540841a5d30c966d113e39636b0545fea9a5ad