jjb: raises quibble db tmpfs from 256M to 320M
mediawiki/extensions/ContentTranslation tests started failling last week
with:
    Error: 1114 The table 'l10n_cache' is full

Selenium tests hit several different languages pages which hit trigger a
localization cache rebuild for that language. The cache is stored in the
database and it eventually overflow the 256MBytes tmpfs we are using for
the database.

With MariaDB 10 on Debian Stretch, just initializing a database already
consumes roughly 110MBytes. Can be tried with:

  $ mkdir /tmp/dummydb
  $ mysql_install_db --datadir=/tmp/dummydb --user="$USER"
  $ du /tmp/dummydb
  8   /tmp/dummydb/performance_schema
  1092    /tmp/dummydb/mysql
  111716  /tmp/dummydb/

Notably due to files:

  | ibdata1 | 12MB
  | ib_logfile0 | 48MB
  | ib_logfile1 | 48MB

While tests are running, the l10n_cache table eventually grow up to
68MBytes and the total size reaches up to 258MBytes. That is too large
for our small tmpfs.

We will need to make Quibble smarter in how it initializes the database
in order to reduce its size.

Bump the database tmpfs from 256MBytes to 320MBytes in the Quibble
docker run commands.

Bug: T217654
Change-Id: Ic02e9554c6997236313cccccbec25faea8e9b4be