jjb: Quibble jobs to capture core files if any
HHVM randomly segfault when running PHPUnit tests and I could not
reproduce on my machine. We thus need to have a core file generated and
capture it.

Our WMCS instances have:

  $ cat /proc/sys/kernel/core_pattern
  /var/tmp/core/core.%h.%e.%p.%t

The Kernel will write to that file in the context of the container (ie
the file is not written on the host). We thus use a Docker volume bound
from the host $WORKSPACE/coredump to /var/tmp/core.

Set ulimit for 'core' at 2GB (2 * 1024^3 B = 2147483648 B).

If a core file file is found, copy it in the host /tmp directory. This
is done using a dummy Docker container to gain root privileges (core
files are only readable by root). We don't want to archive files that
are potentially very large.

This is done before docker-cleanup wipes the workspace entirely.

Bug: T216689
Change-Id: I69d3ec144a89a48ed258ba76ea525ae6c36e53a6