docker: rebuild ci-stretch for debian/libc6 update
Debian has a new 9.8 version with several fix. The parent image has been
rebuild so we have to catch up.

It notably brings up a libc6 update which fix:
  - Fix a use after free in pthread_create().  Closes: #916925.

Which match the stacktraces we get from recently failling builds with
HHVM (T216689).

Rebuild quibble-stretch and quibble-stretch-hhvm. Skip the rest of the
images rebuild since the result is undeterministic.

Bug: T216384
Bug: T216689
Change-Id: I616a7adba8454b04407568702e69a7bcee5bbf2e