fresnel: Add --skip-deps arg to quibble invocation
This skips the 'npm install' step, which is not needed for this
job. Also, that step produces a lot of output due to Fibers used
by wdio, which is nice to avoid having to scroll through.

I did create a "Console Section" pattern that collapses all
preparation steps from Quibble. But, this only works when viewing
"Full Log". The default view linked from Gerrit/Zuul, starts in
the middle of that output, and thus is not matched/hidden..

Change-Id: Ib7d920f91af1c531690833ca68a20885e29d7e41