Update patch set 4
Patch Set 4:

We just noticed during SWAT that CI for a WikibaseLexeme backport needs almost 45 minutes, so if you don’t +2 a backport at the beginning of the SWAT window there’s little chance of merging it. Would this config change have helped with that, reducing the gate-and-submit time?

Patch-set: 4
CC: Lucas Werkmeister (WMDE) <4699@e9e9afe9-4712-486d-8885-f54b72dd1951>