Update patch set 4
Patch Set 4:

Giuseppe and I talked about it this morning for breakfast. That might have a bunch of side effects such as:
* a package that was only available in jessie-backports which would now be unavailable
* anything that relies on an external state that now suddenly brings a new reality / some updates or is broken for some reason. Typically the package managers (apt, npm, composer)

Which is not ideal. An outcome of the conversation we had is that we should have CI to build the containers when a patch is proposed and we need some kind of smoke/end to end tests for those.

Patch-set: 4
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>