Update patch set 1
Patch Set 1:

To build Debian packages, we have a set of instances that use the puppet class packages_builder with all our hooks and setting up cowbuilder images.

We then have the debian-glue job which are based on https://jenkins-debian-glue.org/ . It is really "just" a set of helpers on top of git buildpackage and cowbuilder.  The jobs work fine and are used for the apertium Debian packages.

So my question is: shouldn't we fork sury.org repository to our Gerrit and then add the debian-glue jobs we already have?


Side note: jenkins-debian-glue has some support to publish the build material to an apt repository, though I have actively looked at that option.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>