termbox: build production variant in CI
This is merely an idea, to quickly describe to problem of a breaking
production variant build without noticing during CI.
Only due to the way service-pipeline.groovy is currently implmented
this achieves the desired effect (as soon as we have a helm.yaml this
would also result in a 'Test deployment').
The output could also be somewhat confusing (if it works at all),
running two builds at the same time - a dedicated job running in
parallel would likely be cleaner.

Bug: T220107
Change-Id: I77a3b88b9c0f53eb7ed0d14a6e00432bfec3d98c