Update patch set 2
Patch Set 2:

Potential differences:

* mwext-VisualEditor-docker-publish: Publishes "{src=workspace}/docs" to "doc.wm.o/VisualEditor/{ZUUL_branch}"

* mwext-node10-docs-docker-publish : Publishes src/docs to doc.wm.o/"$DOC_PROJECT/$DOC_SUBPATH/js"

It'll add a subdirectory. Which should be fine, but we'll need to manually wipe what's currently at VisualEditor/master then so that it will not serve the old index.html there that we'll no longer update, but instead become an auto-redirect to the sole subdirectory "js/".

Patch-set: 2
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>