Update patch set 2
Patch Set 2:

Looks fine for the postmerge -publish jobs.

I'm not sure how this is supposed to work for the -patch jobs, since we don't actually generate a full coverage report, and we currently throw them away as soon as we're done diffing (plus its in clover format, not junit).

Depending on how confident you are about this change, you might want to roll it out first to extensions and then MediaWiki core. Extension coverage is trivial and quick to regenerate, while MediaWiki core takes hours and runs on a cron and I'd prefer not to lose coverage info for core.

Patch-set: 2
Reviewer: Legoktm <458@e9e9afe9-4712-486d-8885-f54b72dd1951>