Update patch set 3
Patch Set 3:

The cache is populated by rsync which is run inside a container and uses the 'nobody' user. However the /cache/ directory is owned by root iirc and thus rsync can not update the timestamps on the directory. Hence:

 rsync: failed to set times on "/cache/.": Operation not permitted (1)

That is a non fatal error and can be ignored. It is due to rsync using --archive which implies --times and that option is used to optimize the transfer by comparing timestamps and avoid rsyncing files that have not been changed.  Though, the cache is initially empty so we could probably drop that --times option.  There is also --omit-dir-times


Anyway that is T188488 :)

Patch-set: 3
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>