Update patch set 5
Patch Set 5:

Also, it would be nice to see a project actually generating a job.

Or another thing we could do is that to benefit from Node 10, the project has to migrate to the service pipeline and kubernetes.  This way we don't have to carry all that legacy stuff for the next few years :]

Patch-set: 5
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>