fresnel: Add section anchor to Jenkins links
There is no named anchors unfortunately so this may break in the
future.

It's better to jump to a section that exists and is before the
section we care about (section 3: Fresnel result), then jumping
to a section that doesn't exist or is too far.

Hence, rounding down instead of being exact. There should be
at least 3 sections (fresnel before, fresnel after, fresnel result).

In practice there will also be numerous sections before it (from
castor, quibble, mw-install, …). All of those, as well as fresnel
before and after are collapsed by default. So really linking to
any section that exists and isn't beyond "fresnel result" jumps
to roughly the same spot.

Bug: T220561
Change-Id: I8303fe377fea625b6668bfe46e1a26bca9f56c30