Update patch set 3
Patch Set 3:

extension-gate has jobs such as:

 release-quibble  (branch: REL*)
 wmf-quibble  (branch: master or wmf)

The wmf-quibble branch is for gated extension. It has a hardcoded set of dependencies.

The release-quibble job has dependencies injected from the central CI but only run on REL* branches.

The extension-quibble template is needed to run jobs with dependencies injected from the central CI.


It might make sense to add gate pipelines per branches such as a release pipeline that only reacts to REL branches and a WMF pipeline that only reacts to wmf/* branch. So the config would look like:


 gate-and-submit:  # for master

   # Cover future prod deployment:
   - wmf-quibble-vendor-mysql-hhvm
   - wmf-quibble-vendor-mysql-php72

   # Cover edge case for a future release
   - quibble-php70-postgresql-composer

 release_1_27:
   - release-quibble-php55-sqlite-composer
   ....

 release_1_33:
   - release-quibble-php70-postgresql-composer
   ....

 WMF:  # for wmf/.*
   # The only combo we care about for production
   - npm-node10
   - wmf-quibble-vendor-mysql-hhvm
   - wmf-quibble-vendor-mysql-php72


But I don't even know if we can have several dependent pipelines submitting to the same repos. If that is namespaced by branches, that might be fine.

Patch-set: 3
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>