phan: use Quibble to prepare MediaWiki sources
ci-src-setup-mwext relies on EXT_DEPENDENCIES which should eventually be
phased out. Instead use Quibble to setup clone the extension and prepare
the source directory for a phan run.

Note:
* composer update is run in the extension/skin directory to install
mediawiki/phan-config which carry the phan version to be installed.
* quibble still runs `npm install` and install.php which is suboptimal.

This change will let us phase out the jjb macro docker-ci-src-setup-mwext and remove the legacy setup-mwext.sh script in releng/ci-src-setup.

Bug: T189567
Change-Id: I30bd0b87cc2c679cfbbd27b430247e9e621c44e1