docker: prevent Chromium cache from being saved
The Castor cache for mwselenium-quibble-docker eventually got filled up
due to Chromium caches.
The cache directory would be XDG_CACHE_HOME if the user profile
directory is a child of XDG_CONFIG_HOME which we have set to /tmp in a
parent container.

Relocate XDG_CONFIG_HOME to /tmp/xdg-config-home, thus the Chromium
cache would be in the user profile directory and thus not get saved.

Bug: T220948
Change-Id: Iebadbf8c5b65638159a61f87c6d01bcb8a1ffd94