Update patch set 1
Patch Set 1:

Can we measure how much time it adds to typical gate jobs? Over the past year or so, we've gone from 12 minutes max for the gate pipeline, to now 18-20 minutes in my experience. We really need to think about how to reduce this again. Part of that may be un-doing merging of jobs (e.g. selenium and phpunit), but part is also being stricter about test quality and making sure they can be fast.


I don't know what an upper-bound should be, but I think it's at least worth having the information when making the decision.

Patch-set: 1
CC: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>