Update patch set 1
Patch Set 1:

In this particular case, these tests used to be in gate (inside the Wikibase extension), and dropped out of gate when the code was moved to this extension this week. This is reverting to previous load, not an increase, and so I think we should proceed.

Patch-set: 1
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>