Add MinvervaNeue and Vector to gate
Vector is always cloned by Quibble but surely we should add it to the
gate. MinervaNeue is tightly coupled with MobileFrontend.

I have kept the logic to inject the skin when doing a 'check
experimental' although apparently 'wmf-quibble*' gating jobs are no more
in that pipeline.

Set SKIN_NAME just like we set EXT_NAME.

Bug: T202030
Change-Id: I96d078f43cb1092b66966d27620589692aa6a1f0