beta: limit scap to 512 max file descriptors
When scap runs commands, it closes all potential file descriptors on
each command invocation.  The Jenkins agent on the instance bumps the
maximum number of file descriptors to a million, which takes roughly 240
ms to close.

Those closes pills up to a point where running scap sync takes roughly
12 minutes!

In beta-scap-eqiad job, drop the maximum number of file descriptors to
512 which should be sufficient.

Bug: T204762
Change-Id: I714a1a97b27d54373339a2a7049ccc0d6cb70a20