Update patch set 1
Patch Set 1:

Hmm, now I'm not sure if we want to do this. If we update a dependency version in the extension's composer.json, for example to use a newer version of a library, should it really fail tests until we've changed things in another repo?

Currently, we're only updating our branch of the mediawiki/vendor repo when we're ready to deploy a bunch of changes. With this change, we'd have to update our branch of the mediawiki/vendor repo before testing any changes to the extensions dependencies.

Patch-set: 1
Reviewer: Ejegg <1070@e9e9afe9-4712-486d-8885-f54b72dd1951>