Update patch set 3
Patch Set 3:

(4 comments)

> Thanks for the feedback, hashar. I'll address some of the issues you've pointed out.

To be fair, the only concern I have is the job template having logdir as a parameter.  If one ever wants to override it, it would miss creating the directory. So that is really just a nitpick.

Rest is fine :)


> It's worth mentioning, however, that your comments could very well apply to the majority of the existing code in this project, and it is in need of some serious refactoring and documentation around where to define things, what job template parameters are (un)necessary, code conventions, etc. Without this kind of guidance, contribution is fraught with unexpected rejection and unclear recourse.

that is why i try to give extended code review. But indeed we need a lot of cleanup, doc and refactoring :-(

Patch-set: 3