Update patch set 28
Patch Set 28:

Given that is an experiment we can probably just go with this as is and refactor later before it is generalized.

What I am wondering is whether that sonar job could just be run as part of the test pipeline.  Notably for MediaWiki extensions we use the Zuul template that has a npm test job:

 - name: extension-quibble
   test:
    - mwgate-npm-node-6-docker

So maybe that job can be enhanced to optionally publish to Sonar whenever reports from coverage/checkstyle/junit are found else just skip.

For PHP it is a bit more complicated since "composer test" is run from Quibble. That is done to run the tests (notably parallel-lint) with different versions of PHP.

I don't think we can run coverage in the test pipeline. That would require xdebug / phpdbg to be used which iirc dramatically slow down the tests :/

Patch-set: 28