Update patch set 12
Patch Set 12:

(2 comments)

My only concern is to have the Docker image to match Quibble version which is currently 0.0.31. So the container should be 0.0.31-3  not 0.0.32 :)

I usually only bump the job that really need to be bumped, ie just the coverage job. The reason is that the container is rebuild from scratch and could well come up with new package version which might cause some jobs to fail :/

So I would recommend to only bump the mwext-phpunit-coverage* jobs and leave the others untouched.  We can bump them later on in a different change.  But I am maybe too paranoid.

Regardless +1 on having junit files generated when doing the coverage report :]

Patch-set: 12
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>