Update patch set 2
Patch Set 2: Code-Review-1

The '-m checkstyle' should be a different change. The issue is that 1.2.x no more allows options after files and I believe that since 1.2.3 and commit:

https://github.com/phan/phan/commit/5d571563166904a724f24b11c1e669208e495a75


Instead of hardcoding '-m checkstyle' in the entry point, we should just adjust the phan invocation by the Jenkins job:

 dockerfiles/mediawiki-phan/README.md:16:    /mediawiki/extensions/ElectronPdfService -m checkstyle
 jjb/mediawiki-extensions.yaml:110:        args: '/$ZUUL_PROJECT -m checkstyle'


 - args: '/$ZUUL_PROJECT -m checkstyle'
 + args: '/-m checkstyle /$ZUUL_PROJECT'



The reason I am picky on this one, is that we also would have to change the job to write the checkstyle file to a specific file (eg $LOG_DIR/junit-phan.xml)  since currently lot of jobs do not even have the checkstyle report processed. But that is a different story.




The COLUMNS hack should be added to run-libraries.sh as well :]

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1