docker: disable phan progress bar
`phan -p` displays a progress bar which is generated via
Symfony\Component\Console\Terminal, to determine the width of the
terminal, the library shells out to stty which causes a fork. When under
memory pressure, the Linux kernel would reject the fork assuming the
child process might end up using as much memory as the parent process.

That can be fixed by instructing the kernel to honor the memory
allocation by setting vm.overcommit_memory = 1, or to just preventing
the fork in the first place.

Hence, in the mediawiki-phan entry points, remove '-p' to disable the
progress bar and thus the fork() call.

Details:
https://phabricator.wikimedia.org/T219114#5084302

Bug: T219114
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Id74ae0ec2beebe0ad13b034fcf9d072f9ee593aa