Remove php5/7.0/7.1 from gate-and-submit-swat
Changes only enter gate-and-submit-swat pipeline when a +2 is for a wmf
branch. Nowadays Wikimedia production only needs HHVM/php72.

Remove php55/php70/php71 jobs when there are php70/hhvm jobs.
Remove release-* jobs, they are intended for REL_ branches.

Add some FIXME for later investigation.

We can most probably remove a few skip-if rules on the jobs, but that
can be done in a later change.

Bug: T221410
Change-Id: I2fe51c1024670d6a9c15c141a49b370628667e45