Update patch set 1
Patch Set 1:

Since gate-and-submit-swat is only for wmf/ branches, we can just remove the jobs. That saves us from having to deal with the complicated skip-if rules that override each others.

Probably we could generalize that and add DependentPipelines for each of the REL branches. Then I guess we can drop some skip-if rules and instead assert that a job is only in the proper pipelines.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>