Update patch set 1
Patch Set 1:

> Hmm, now I'm not sure if we want to do this. If we update a
 > dependency version in the extension's composer.json, for example to
 > use a newer version of a library, should it really fail tests until
 > we've changed things in another repo?

This seems correct to me, since the production deployment will be running vendor.git.  You'll need to deploy `vendor` before any DonationInterface code relying on the new library can be deployed, so it's reasonable that the CI tests fail until the precondition is satisfied...

 > 
 > Currently, we're only updating our branch of the mediawiki/vendor
 > repo when we're ready to deploy a bunch of changes. With this
 > change, we'd have to update our branch of the mediawiki/vendor repo
 > before testing any changes to the extensions dependencies.

Oh noes!  That makes sense, too.  One solution might be to merge the vendor change to a "staging" branch, then prepare the DI changes against that branch.  That should trigger Zuul logic to checkout vendor on a matching branch!  Once the patches are merged to staging on both vendor and CI, then you can prepare the vendor merge to master (DI code must be forwards-compatible, of course), and finally a DI merge to master.

Patch-set: 1
CC: Awight <143@e9e9afe9-4712-486d-8885-f54b72dd1951>