docker: mediawiki phan wrapper requires a path
The MediaWiki phan wrapper requires the extension path to be the first
argument. However when not using the wrapper, that causes phan to abort
since it requires options first and then the files/paths to check.

A previous upgraded has show Wikibase to fail with:

 + exec /srv/phan/vendor/bin/phan -d . -p \
    /mediawiki/extensions/Popups \
    -m checkstyle
 Unknown option '-m'
 Usage: /srv/phan/vendor/bin/phan [options] [files...]

https://phabricator.wikimedia.org/T219114#5176487

Always inject the path when using Phan 0.8/mediawiki wrapper. This is
only for repositories using phan 0.8 and the MediaWiki wrappers which
turns out to be only mediawiki/extensions/Wikibase

Bug: T219114
Change-Id: I5c26f0e75a0296e762617a1ac1f047d5e11cee58