phan: do not install mediawiki dev dependencies
When running Phan for an extension/skins, we had Quibble installing
mediawiki core dev dependencies which could overlap with the one from
the extension/skins and would cause an error due to duplicate interfaces
(PhanRedefinedInheritedInterface).

The reason is that when I have migrated to use Quibble (66e04e1e), I
have missed the fact that the old script was not install MediaWiki
development dependencies (composer update --no-dev).

* Rename JJB macro to docker-mwext-for-phan since it is only used
  for Phan preparation.
* Switch Quibble to use vendor.git (which partially revert cc6e0371
  which switched to composer) and use --skip-deps to prevent it from
  running composer update entirely.
* Add a step to run 'composer update --no-dev' for the extension/skin.

Bug: T223397
Change-Id: If7171a6d46473b0bb05e3adaeb40229e7881c2b6