Update patch set 1
Patch Set 1: Code-Review+1

But won't this mean that we're not testing the released versions of /vendor? In the long-term I understand the idea is to stop having released versions of /vendor, but until that date, not having a testing strategy isn't ideal.

Patch-set: 1
Reviewer: Jforrester <366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1