Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2:
> 
> (2 comments)
> 
> I am not sure what was wrong in T219749
> 
> 
> The wikidata-query-gui-npm-browser-node-6-docker job being run forwikidata/query/gui should be able to run the selenium tests just fine. It is probably worth adding the wdio task back so that they get run when someone does "npm test".
> 
> Imho it should just work.
> 
> 
> 
> The issue Stas encountered on T219749 is different.  The failure comes from the CI tests for wikidata/query/rdf which are run by the Maven test suite.  The job doing that is wikidata-query-rdf-maven-java8-docker which really just ship java8 and then runs maven.  There is no chromedriver being spawned.
> 
> The wikidata/query/rdf test suite then run the GUI test suite using npm, but since the gui code uses wdio/chromedriver and the maven image does not have it ... The suite fails to connect to chromedriver ( ECONNREFUSED 127.0.0.1:4444 ).
> 
> 
> I am not sure whether the GUI test should be run as part of testing wikidata/query/rdf .  Probably only a subset of them should?

That was exactly the issue we faced when we first tried to run wdio as part of test. I think rdf should not be running gui tests to begin with. So for that reason, I would not go ahead and add browser tests into the image for the rdf one and prefer, for now, to run browser tests only in gui only.

Patch-set: 2
Reviewer: Alaa Sarhan <6768@e9e9afe9-4712-486d-8885-f54b72dd1951>