Publish coverage reports from extension-codehealth
By doing this, we can safely drop extension-coverage from the extensions which
have extension-codehealth enabled, since:

1. SonarQube calculates coverage on a per-patch basis
2. On postmerge, we take the full coverage report and publish it with the builder

The end result is that we'll eliminate redundant code coverage generation
reports for extensions in extension-codehealth.

Caveat: I have no idea if adding a builder to the
mwext-codehealth-master-non-voting job is possible to do as I've implemented in
this patch, nor have I tested this.

Change-Id: Ic1d5ac41cf29bb5d990ce1a3074d9f644a00614b