Update patch set 1
Patch Set 1:

CI wise yes that is all that is needed. As for skipping slow tests or tests that are not relevant, there is no system for it yet :-( CirrusSearch add a minute and a half of test, and Scribunto is not added because it adds a couple minutes.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>