Update patch set 2
Patch Set 2:

> mwselenium-legacy was already removed; I thought that mwselenium-quibble-docker was the new one? Still, it's not working yet, so removing for now.

The first, mwselenium-legacy, is a Zuul template which does:

  - name: mwselenium-legacy
    test:
      - mwselenium-quibble-docker
    postmerge:
      - mwselenium-quibble-docker

The second is the Jenkins job that runs the ruby based testsuite using Quibble to set up MediaWiki (yeah I know, lot of layers).


mwselenium-quibble-docker is not in gate-and-submit because it is rather slow on some repositories hence why the template has it in postmerge instead.

Then eventually the job also get added to gate-and-submit on a per repo basis.


;D

Patch-set: 2
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>