Update patch set 1
Patch Set 1:

LGTM, but I'd do it in two steps: first, I'd put these as experimental jobs and move them to voting in a second step. We are not in a hurry to have the images built, so ensuring everything works as expected using non-voting jobs could be a good idea. That said, I'm not opposed to moving them to voting jobs straightway.

Patch-set: 1
Reviewer: Mobrovac <2177@e9e9afe9-4712-486d-8885-f54b72dd1951>