Update patch set 1
Patch Set 1:

We also need to get rid of the Jenkins jobs. They are defined in the file jjb/mediawiki-services.yaml with the block:

 - project:
     name: recommendation-api
     jobs:
      - '{name}-npm-node-6-docker'
      - '{name}-deploy-npm-node-6-docker'


The second one, '{name}-deploy-npm-node-6-docker', is for the /deploy repo which I guess we will archive.  That can be done in zuul/layout.yaml by replacing:


  - name : mediawiki/services/recommendation-api/deploy
    test:
      - recommendation-api-deploy-npm-node-6-docker
    gate-and-submit:
      - recommendation-api-deploy-npm-node-6-docker

With:

  - name : mediawiki/services/recommendation-api/deploy
    template:
      - name: archived

Then we mark the repository read only in Gerrit.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>