Update patch set 1
Patch Set 1:

The skip-if filters are a bit of a mess right now.  The repository has the Zuul template extension-phan which currently still uses php7.0.  There is an effort to migrate the phan job to PHP 7.2 but it is not complete yet.

So instead, I am going to send a new change that just remove phan entirely for this repo, and eventually we can add it back once the PHP 7.2 job is ready.


For the other skip-if and dropping HHVM. I have to look into it a bit more carefully.

Patch-set: 1
Reviewer: Hashar <24@e9e9afe9-4712-486d-8885-f54b72dd1951>