Ensure test has proper coverage
test_mwcore_switch_to_quibble() ensure that jobs for
test/gate-and-submit of mediawiki/core are or are not triggered. Thus
the test data has to explicitly list them all.

A couple jobs got removed and the test did not get updated. Add an
assertion to ensure the test data are in sync with the job definition.

Change-Id: I0a0b5a18166ed6616da2bb7b8b1a0ae95bfd7f6c