Maintenance: cleanup contint1001, too
Adds 'contint1001' to the list of machines we clean every 5 minutes if
they hit a certain threshold.

This skips taking contint1001 offline if the disk usage is above 90,
since that is unnecessary in this case; that is, it is likely that there
is some other cause for the disk being full other than jobs being run on
this node (either docker-pkg runs or jenkins logs).

Slightly refactors this maintenance script to allow for machines where
the docker partition is not "/var/lib/docker".

docker-pkg tags each image it builds with "latest". This script will
find all images, using their image id as a key and their tags as a list
referenced in a map, e.g.,

  29397cdce9f7: ['latest', '20190624']
  faec7ee79468: ['20190624']

The script then loops through the map and removes images that are not
tagged 'latest'.

In the case that we are using docker-pkg to build from a version of an
image that is not also the latest version of that image, subsequent
docker-pkg runs will pull down the missing image (since on contint1001 a
registry is defined in the docker-pkg config). This will take longer,
but is preferable to contint1001 running out of disk space.

Bug: T207702
Change-Id: I6d46e9f901d02dafe3933f57559d7bd6d78e6b7b