docker: Install php-ast 1.0.0 for seccheck, back to PHP7.0
Install php-ast1.0.0 together with 0.1.2, so that this job can run both
seccheck <2.0.0 and >= 2.0.0. Moreover, switch back to PHP7.0, to ensure
that we won't have compatibility issues.
Note that this patch doesn't use the php-ast container because we need
0.1.2, which isn't there anyway. This is intended to be a temporary
change. When we switch to PHP72, we'll only have to keep php-ast 1.0.1,
possibly taken from the ast container, and some other code will be
simplifiable.

Some of the hacks included here are taken from the mediawiki-phan
container.

Note: Did not test this locally (yet).

Bug: T227385
Change-Id: I8dae352d4997b9df33169d1b13c371f5e75bc29b