jjb: [mwext-codehealth*] Only clone ZUUL_PROJECT
For the codehealth jobs, we don't need to clone all dependencies of an extension
since the unit tests can run without these dependencies. This should speed up
the builds a bit, both for cloning and composer install steps.

Use zuul-cloner to get mediawiki-core and the extension. Then we can pass
--skip-zuul to quibble. We are *not* passing --skip-deps because we need both
composer and npm to install.

Change-Id: I425b121fcb60588fff6ccceb31cc85ca3c6d0888