zuul: pipelines for fundraising branches
mediawiki/core@fundraising/REL1_31 is tests with a very specific
integration job that runs with DonationInterface:

 quibble-donationinterface-REL1_31-php70-docker

Since it is in the test and gate-and-submit pipeline, we have to set
branch filters for all the others jobs to skip them on fundraising and
only retain that specific job.

Instead add two new pipelines (test-fundraising and
gate-and-submit-fundraising) that only react to events to the
fundraising branch.  That dramatically simplify the configuration at the
expense of adding a couple more pipelines.

Have test and gate-and-submit events to filter out the fundraising
branches.
Move the quibble-donation job to the new pipelines.
There is no more need to skip-if the job for non fundraising/REL
branches since it is no more in the other pipelines.

A future change can then cleanup the fundraising/REL branch filters. We
would most probably want to have such pipeline arranged for each of the
release branches we maintain.

Change-Id: I7f03ced4acab67276e97b2b2498a8a9b82c85328