zuul: drop skip-if for fundraising/REL branch
Now that fundraising/REL events are solely processed in test-fundraising
and gate-and-submit-fundraising, we no more have to apply a skip-if
filters to jobs not in those pipelines.

Remove the branch filter from all quibble jobs and
mediawiki-core-php72-phan-docker since none are triggered in the
fundraising pipelines anyway.

Drop tests that no more apply since they are never associated with the
fundraising pipelines.

Change-Id: I8173c1ab51efe24be0dcd4eb3d1c16d86874faa4