zuul: introduce test-wmf pipeline
Same as the 'test' pipeline but solely for the wmf/* branches.

Add a branch filter to all events of the 'test' pipeline to skip events
made to wmf/* branches in the same way we did for gate-and-submit-wmf.

For all projects, copy the 'test' definition to 'test-wmf' whenever
there is already a gate-and-submit-wmf pipeline defined for the project.

Add a test to ensure projects either have both test-wmf and
gate-and-submit-wmf or that they have none of them.
Ensure that an event to wmf branch is rejected by test but accepted by
test-wmf.

Effectively the test pipeline definition is duplicated to test-wmf
whenever gate-and-submit-wmf is applied to the project. As follow up,
we can then remove lot of the annoying skip-if filter we had set on
jobs.

Change-Id: I880333fab06a2054c7a9a47a2f8deae3d2133181