Apply REL based pipelines
Copy paste from test/gate-and-submit to each of the pair of pipelines
dedicated to REL1_31, REL1_32 and REL1_33. Some adjustements have been
made such as:

* Default to use php70 instead of HHVM in the test pipeline
* Never use vendor (broken on release branches) (T189560).

Remove release-quibble jobs, their original intent was to apply skip-if
filters on them to select what we wanted to run on REL branches. It is
much easier with the new dedicated pipelines.

Remove most of the filters.
Remove bunch of tests for the jobs filters.

Bug: T189560
Change-Id: I4ef26800cb8c52fdaea1b02da627c0c36e4aeb60